# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI-Native Observability & Threat Detection Engine for Global Payment APIs",
    "description": "Enterprise-grade runtime security, prompt injection detection, financial logic validation, and PCI DLP redaction engine for autonomous payment agents.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/security/scan": {
      "post": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "Full Multi-Layer Security Scan",
        "description": "Performs full zero-trust security scan combining prompt injection,\njailbreak, financial manipulation, logic validator, and DLP layers.",
        "operationId": "scan_full_api_v1_security_scan_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FullSecurityScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityScanResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/scan-prompt": {
      "post": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "Dedicated Prompt Threat Scan",
        "description": "Fast-path scan focusing specifically on user prompt injection,\njailbreak attempts, and financial instruction manipulation.",
        "operationId": "scan_prompt_api_v1_security_scan_prompt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PromptScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/scan-response": {
      "post": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "DLP and PAN Response Scanner",
        "description": "Outbound response DLP scanner. Detects, validates with Luhn,\nand masks PANs, CVVs, and credentials before user delivery.",
        "operationId": "scan_response_api_v1_security_scan_response_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResponseScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/events": {
      "get": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "List Security Events",
        "description": "Lists security events with optional severity filter.",
        "operationId": "list_security_events_api_v1_security_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by severity: LOW, MEDIUM, HIGH, CRITICAL",
              "title": "Severity"
            },
            "description": "Filter by severity: LOW, MEDIUM, HIGH, CRITICAL"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityEventListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/security/events/{event_id}": {
      "get": {
        "tags": [
          "Security & Threat Detection"
        ],
        "summary": "Get Security Event by ID",
        "description": "Retrieves full details of a specific security event.",
        "operationId": "get_security_event_api_v1_security_events__event_id__get",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Event Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SecurityEventResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts": {
      "get": {
        "tags": [
          "Alerts & Incident Response"
        ],
        "summary": "List Security Alerts",
        "description": "Lists alerts with active count and resolution status filtering.",
        "operationId": "list_alerts_api_v1_alerts_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "is_resolved",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by resolution status",
              "title": "Is Resolved"
            },
            "description": "Filter by resolution status"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/alerts/{alert_id}/resolve": {
      "post": {
        "tags": [
          "Alerts & Incident Response"
        ],
        "summary": "Resolve Alert",
        "description": "Marks an active security alert as investigated and resolved by an analyst.",
        "operationId": "resolve_alert_api_v1_alerts__alert_id__resolve_post",
        "parameters": [
          {
            "name": "alert_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Alert Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AlertResolveRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AlertResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AlertListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "active_count": {
            "type": "integer",
            "title": "Active Count"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/AlertResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "active_count",
          "items"
        ],
        "title": "AlertListResponse"
      },
      "AlertResolveRequest": {
        "properties": {
          "resolved_by": {
            "type": "string",
            "title": "Resolved By"
          },
          "resolution_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Notes"
          }
        },
        "type": "object",
        "required": [
          "resolved_by"
        ],
        "title": "AlertResolveRequest"
      },
      "AlertResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "security_event_id": {
            "type": "string",
            "title": "Security Event Id"
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "is_resolved": {
            "type": "boolean",
            "title": "Is Resolved"
          },
          "resolved_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved At"
          },
          "resolved_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved By"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "security_event_id",
          "severity",
          "title",
          "description",
          "is_resolved",
          "created_at"
        ],
        "title": "AlertResponse"
      },
...
      "FullSecurityScanRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt"
          },
          "proposed_transaction": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Transaction"
          },
          "agent_response": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Response"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "default": "usr_client"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "FullSecurityScanRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PromptScanRequest": {
        "properties": {
          "prompt": {
            "type": "string",
            "title": "Prompt",
            "description": "Prompt to scan for injection and jailbreak"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "default": "usr_unknown"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          }
        },
        "type": "object",
        "required": [
          "prompt"
        ],
        "title": "PromptScanRequest"
      },
      "ResponseScanRequest": {
        "properties": {
          "response_text": {
            "type": "string",
            "title": "Response Text",
            "description": "Agent response to scan for PAN and sensitive data"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "agent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Agent Id",
            "default": "autonomous_payment_agent_v1"
          }
        },
        "type": "object",
        "required": [
          "response_text"
        ],
        "title": "ResponseScanRequest"
      },
      "RiskSignal": {
        "properties": {
          "detector": {
            "type": "string",
            "title": "Detector"
          },
          "is_threat": {
            "type": "boolean",
            "title": "Is Threat"
          },
          "risk_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score"
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Severity"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "technique": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Technique"
          },
          "targeted_control": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Targeted Control"
          },
          "explanation": {
            "type": "string",
            "title": "Explanation"
          },
          "metadata": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "detector",
          "is_threat",
          "risk_score",
          "severity",
          "confidence",
          "explanation"
        ],
        "title": "RiskSignal"
      },
      "SecurityEventListResponse": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/SecurityEventResponse"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "total",
          "items"
        ],
        "title": "SecurityEventListResponse"
      },
      "SecurityEventResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transaction Id"
          },
          "agent_id": {
            "type": "string",
            "title": "Agent Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "severity": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Severity"
          },
          "risk_score": {
            "type": "number",
            "title": "Risk Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "FLAG",
              "BLOCK",
              "REDACT_AND_ALLOW"
            ],
            "title": "Action"
          },
          "payload_snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Snippet"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "agent_id",
          "event_type",
          "severity",
          "risk_score",
          "confidence",
          "source",
          "description",
          "action",
          "created_at"
        ],
        "title": "SecurityEventResponse"
      },
      "SecurityScanResponse": {
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "title": "Timestamp"
          },
          "unified_assessment": {
            "$ref": "#/components/schemas/UnifiedRiskAssessment"
          },
          "security_event_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Security Event Ids"
          }
        },
        "type": "object",
        "required": [
          "timestamp",
          "unified_assessment"
        ],
        "title": "SecurityScanResponse"
      },
...
      "UnifiedRiskAssessment": {
        "properties": {
          "risk_score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Risk Score",
            "description": "Calibrated unified risk score (0-100)"
          },
          "risk_level": {
            "type": "string",
            "enum": [
              "LOW",
              "MEDIUM",
              "HIGH",
              "CRITICAL"
            ],
            "title": "Risk Level"
          },
          "recommended_action": {
            "type": "string",
            "enum": [
              "ALLOW",
              "FLAG",
              "BLOCK",
              "REDACT_AND_ALLOW"
            ],
            "title": "Recommended Action"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/RiskSignal"
            },
            "type": "array",
            "title": "Signals"
          },
          "masked_output": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Masked Output"
          },
          "summary_explanation": {
            "type": "string",
            "title": "Summary Explanation"
          }
        },
        "type": "object",
        "required": [
          "risk_score",
          "risk_level",
          "recommended_action",
          "signals",
          "summary_explanation"
        ],
        "title": "UnifiedRiskAssessment"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}